refactor-cm: Create single source for codemirror config such that not double maintenance required
