+++ app/src/views/items/EditItem.vue
@@ -76,21 +76,16 @@
 
 <script>
 import * as bootstrap from "bootstrap";
-
-import * as CodeMirror from "codemirror";
-import { linter, lintGutter } from "@codemirror/lint";
-import { oneDark } from "@codemirror/theme-one-dark";
-import { json, jsonParseLinter } from "@codemirror/lang-json";
-
 import { useRoute, useRouter } from "vue-router";
-import { computed, inject, onMounted, reactive, ref, watch } from "vue";
+import { inject, onMounted, reactive, ref, watch } from "vue";
 
 import { getTable } from "@/services/table";
+import codeMirrorConfig from "./codeMirrorConfig";
 import { getItem, updateItem } from "@/services/item";
 
 export default {
   setup() {
-    let cm;
+    let codeMirror;
     const item = ref("");
     const editItem = ref("");
 
@@ -167,34 +162,15 @@ export default {
 
     // https://codemirror.net/
     onMounted(async () => {
-      const Theme = CodeMirror.EditorView.theme({
-        ".cm-line, .cm-line span": {
-          "font-size": "13px",
-          "font-weight": "bold",
-          "font-family": "'Fira Code', monospace !important",
-        },
-        ".cm-tooltip": {
-          "background-color": "#fafafa",
-        },
-      });
-
-      cm = editorFromTextArea(textAreaRef.value, [
-        CodeMirror.basicSetup,
-        json(),
-        linter(jsonParseLinter()),
-        lintGutter(),
-        Theme,
-        oneDark,
-        // oneDarkHighlightStyle,
-        CodeMirror.EditorView.updateListener.of((update) => {
-          editItem.value = update.state.doc.toString();
-        }),
-      ]);
-
       await initItem();
+      codeMirror = codeMirrorConfig(textAreaRef, editItem);
 
-      cm.dispatch({
-        changes: { from: 0, to: cm.state.doc.length, insert: item.value },
+      codeMirror.dispatch({
+        changes: {
+          from: 0,
+          to: codeMirror.state.doc.length,
+          insert: item.value,
+        },
       });
     });
 
@@ -210,17 +186,6 @@ export default {
 
       item.value = JSON.stringify(Item, null, 2);
       store.table.setters.setTable(table);
-    };
-
-    const editorFromTextArea = (textarea, extensions) => {
-      const view = new CodeMirror.EditorView({
-        doc: JSON.stringify("", null, "\t"),
-        extensions,
-      });
-      textarea.parentNode.insertBefore(view.dom, textarea);
-      textarea.style.display = "none";
-
-      return view;
     };
 
     return {
